refactor: update command naming
